removed dynamic import on entire page contents as it turned out to be unnecessary
