+++ project-base/storefront/pages/index.tsx
@@ -1,4 +1,13 @@
+import { SearchMetadata } from 'components/Basic/Head/SearchMetadata';
+import { Banners } from 'components/Blocks/Banners/Banners';
+import { DeferredBlogPreview } from 'components/Blocks/BlogPreview/DeferredBlogPreview';
+import { PromotedCategories } from 'components/Blocks/Categories/PromotedCategories';
+import { DeferredPromotedProducts } from 'components/Blocks/Product/DeferredPromotedProducts';
+import { DeferredLastVisitedProducts } from 'components/Blocks/Product/LastVisitedProducts/DeferredLastVisitedProducts';
+import { CommonLayout } from 'components/Layout/CommonLayout';
 import { PageDefer } from 'components/Layout/PageDefer';
+import { Webline } from 'components/Layout/Webline/Webline';
+import { useDomainConfig } from 'components/providers/DomainConfigProvider';
 import { BLOG_PREVIEW_VARIABLES } from 'config/constants';
 import {
     TypeBlogArticlesQueryVariables,
@@ -13,19 +22,59 @@ import {
 } from 'graphql/requests/products/queries/RecommendedProductsQuery.generated';
 import { SliderItemsQueryDocument } from 'graphql/requests/sliderItems/queries/SliderItemsQuery.generated';
 import { TypeRecommendationType } from 'graphql/types';
+import { GtmPageType } from 'gtm/enums/GtmPageType';
+import { useGtmStaticPageViewEvent } from 'gtm/factories/useGtmStaticPageViewEvent';
+import { useGtmPageViewEvent } from 'gtm/utils/pageViewEvents/useGtmPageViewEvent';
 import { NextPage } from 'next';
+import useTranslation from 'next-translate/useTranslation';
 import dynamic from 'next/dynamic';
 import { getServerSidePropsWrapper } from 'utils/serverSide/getServerSidePropsWrapper';
 import { ServerSidePropsType, initServerSideProps } from 'utils/serverSide/initServerSideProps';
 
-const HomePageContent = dynamic(() =>
-    import('components/Pages/HomePage/HomePageContent').then((component) => component.HomePageContent),
+const RecommendedProducts = dynamic(() =>
+    import('components/Blocks/Product/RecommendedProducts').then((component) => component.RecommendedProducts),
 );
 
 const HomePage: NextPage<ServerSidePropsType> = () => {
+    const { t } = useTranslation();
+    const { isLuigisBoxActive } = useDomainConfig();
+
+    const gtmStaticPageViewEvent = useGtmStaticPageViewEvent(GtmPageType.homepage);
+    useGtmPageViewEvent(gtmStaticPageViewEvent);
+
     return (
         <PageDefer>
-            <HomePageContent />
+            <SearchMetadata />
+            <CommonLayout>
+                <Webline className="mb-14">
+                    <Banners />
+                </Webline>
+
+                <Webline className="mb-6">
+                    <h2 className="mb-3">{t('Promoted categories')}</h2>
+                    <PromotedCategories />
+                </Webline>
+                {isLuigisBoxActive && (
+                    <RecommendedProducts
+                        recommendationType={TypeRecommendationType.Personalized}
+                        render={(recommendedProductsContent) => (
+                            <Webline className="mb-6">
+                                <h2 className="mb-3">{t('Recommended for you')}</h2> {recommendedProductsContent}
+                            </Webline>
+                        )}
+                    />
+                )}
+                <Webline className="mb-6">
+                    <h2 className="mb-3">{t('Promoted products')}</h2>
+                    <DeferredPromotedProducts />
+                </Webline>
+
+                <Webline type="blog">
+                    <DeferredBlogPreview />
+                </Webline>
+
+                <DeferredLastVisitedProducts />
+            </CommonLayout>
         </PageDefer>
     );
 };
