+++ project-base/storefront/pages/products/[productSlug].tsx
@@ -1,31 +1,86 @@
+import { DeferredLastVisitedProducts } from 'components/Blocks/Product/LastVisitedProducts/DeferredLastVisitedProducts';
+import { SkeletonPageProductDetail } from 'components/Blocks/Skeleton/SkeletonPageProductDetail';
+import { SkeletonPageProductDetailMainVariant } from 'components/Blocks/Skeleton/SkeletonPageProductDetailMainVariant';
+import { CommonLayout } from 'components/Layout/CommonLayout';
 import { PageDefer } from 'components/Layout/PageDefer';
 import {
     ProductDetailQueryDocument,
     TypeProductDetailQuery,
     TypeProductDetailQueryVariables,
+    useProductDetailQuery,
 } from 'graphql/requests/products/queries/ProductDetailQuery.generated';
 import {
     TypeRecommendedProductsQueryVariables,
     RecommendedProductsQueryDocument,
 } from 'graphql/requests/products/queries/RecommendedProductsQuery.generated';
 import { TypeRecommendationType } from 'graphql/types';
+import { useGtmFriendlyPageViewEvent } from 'gtm/factories/useGtmFriendlyPageViewEvent';
+import { useGtmPageViewEvent } from 'gtm/utils/pageViewEvents/useGtmPageViewEvent';
 import { NextPage } from 'next';
 import dynamic from 'next/dynamic';
+import { useRouter } from 'next/router';
 import { OperationResult } from 'urql';
 import { createClient } from 'urql/createClient';
 import { handleServerSideErrorResponseForFriendlyUrls } from 'utils/errors/handleServerSideErrorResponseForFriendlyUrls';
 import { getSlugFromServerSideUrl } from 'utils/parsing/getSlugFromServerSideUrl';
+import { getSlugFromUrl } from 'utils/parsing/getSlugFromUrl';
 import { getServerSidePropsWrapper } from 'utils/serverSide/getServerSidePropsWrapper';
 import { ServerSidePropsType, initServerSideProps } from 'utils/serverSide/initServerSideProps';
 
-const ProductDetailWrapper = dynamic(() =>
-    import('components/Pages/ProductDetail/ProductDetailWrapper').then((component) => component.ProductDetailWrapper),
+const ProductDetailContent = dynamic(
+    () =>
+        import('components/Pages/ProductDetail/ProductDetailContent').then(
+            (component) => component.ProductDetailContent,
+        ),
+    {
+        loading: () => <SkeletonPageProductDetail />,
+    },
 );
 
+const ProductDetailMainVariantContent = dynamic(
+    () =>
+        import('components/Pages/ProductDetail/ProductDetailMainVariantContent').then(
+            (component) => component.ProductDetailMainVariantContent,
+        ),
+    {
+        loading: () => <SkeletonPageProductDetailMainVariant />,
+    },
+);
 const ProductDetailPage: NextPage<ServerSidePropsType> = () => {
+    const router = useRouter();
+    const [{ data: productData, fetching }] = useProductDetailQuery({
+        variables: { urlSlug: getSlugFromUrl(router.asPath) },
+    });
+
+    const product =
+        productData?.product?.__typename === 'RegularProduct' || productData?.product?.__typename === 'MainVariant'
+            ? productData.product
+            : null;
+
+    const pageViewEvent = useGtmFriendlyPageViewEvent(product);
+    useGtmPageViewEvent(pageViewEvent, fetching);
+
     return (
         <PageDefer>
-            <ProductDetailWrapper />
+            <CommonLayout
+                breadcrumbs={product?.breadcrumb}
+                breadcrumbsType="category"
+                canonicalQueryParams={[]}
+                description={product?.seoMetaDescription}
+                hreflangLinks={product?.hreflangLinks}
+                isFetchingData={fetching}
+                title={product?.seoTitle || product?.name}
+            >
+                {product?.__typename === 'RegularProduct' && (
+                    <ProductDetailContent fetching={fetching} product={product} />
+                )}
+
+                {product?.__typename === 'MainVariant' && (
+                    <ProductDetailMainVariantContent fetching={fetching} product={product} />
+                )}
+
+                <DeferredLastVisitedProducts currentProductCatnum={product?.catalogNumber} />
+            </CommonLayout>
         </PageDefer>
     );
 };
